+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2020-07-08
+### Added
+- `setup.sh` to install all package requirements
+- methods for querying the nouns file (see `german_nouns/query`)
+
+### Changed
+- directory structure
+- README.md
+
 ## [1.0.1] - 2020-07-06
 ### Added
 - [wiktionary_de_parser](https://github.com/gambolputty/wiktionary_de_parser) requirement